No need to curry rule that was successful
We don't rely on success result data in any way so it's safe to just
return original predicate. In the future we should make detailed results
completely optional so that ie dry-validation could ask only for errors
so that we are not going to create any heavy success result objects when
we don't need them.